feat(profile): refactor profile page to utilize workout session data and remove calendar component chore(workout-sessions): delete unused workout sessions page feat(header): add home icon with tooltip to header for better navigation chore(calendar): remove calendar component as it is no longer used fix(release-notes): add tooltip to release notes button for improved accessibility refactor(workout-builder): streamline workout session handling and remove unused code in workout stepper and session sets components
