+++ src/features/release-notes/ui/release-notes-dialog.tsx
@@ -5,6 +5,7 @@ import { Bell } from "lucide-react";
 
 import { useCurrentLocale, useI18n } from "locales/client";
 import { formatDate } from "@/shared/lib/date";
+import { InlineTooltip } from "@/components/ui/tooltip";
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogTrigger } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
 
@@ -18,8 +19,9 @@ export function ReleaseNotesDialog() {
     <Dialog>
       <DialogTrigger asChild>
         <Button aria-label={t("release_notes.release_notes")} className="rounded-full hover:bg-slate-100" size="small" variant="ghost">
-          <span className="sr-only">{t("release_notes.release_notes")}</span>
-          <Bell className="text-slate-700 h-5 w-5" />
+          <InlineTooltip title="Annonces / Changelog">
+            <Bell className="text-blue-500 h-6 w-6" />
+          </InlineTooltip>
         </Button>
       </DialogTrigger>
       <DialogContent className="max-w-md">
